DZ-4

+++ models/contacts.js
@@ -1,26 +1,26 @@
 const Contact = require("../service/contactSchema");
 
-const listContacts = async () => {
-  return await Contact.find()
+const listContacts = async (_id) => {
+  return await Contact.find({owner: _id },{__v: 0, owner: 0})
 }
-const getContactById = async (contactId) => {
- return Contact.findOne({ _id: contactId })
-}
-
-const removeContact = async (contactId) => {
-  return Contact.findByIdAndRemove({ _id: contactId })
+const getContactById = async (contactId, id) => {
+ return Contact.findOne({ _id: contactId, owner: id}, {__v: 0, owner: 0})
 }
 
-const addContact = async ({name, email, phone}) => {
-  return Contact.create({ name, email, phone })
+const removeContact = async (contactId, id) => {
+  return Contact.findByIdAndRemove({ _id: contactId , owner: id}, {__v: 0, owner: 0})
 }
 
-const updateContact = async (contactId, body) => {
-  return Contact.findByIdAndUpdate({ _id: contactId }, body, { new: true })
+const addContact = async ({name, email, phone}, owner) => {
+  return Contact.create({ name, email, phone, owner })
 }
-const updateStatusContact = async (favorite, contactId) => {
+
+const updateContact = async (contactId, body, id) => {
+  return Contact.findByIdAndUpdate({ _id: contactId, owner: id }, body, { new: true })
+}
+const updateStatusContact = async (favorite, contactId, id) => {
   return await Contact.findByIdAndUpdate(
-      contactId, { favorite }, { new: true })
+      {_id:contactId, owner: id}, { favorite }, { new: true })
 }
 
 module.exports = {
